Fix handling of union initializers in PredicateCPA if cpa.predicate.handleImplicitInitialization is true (which is the default).
The coded added initializing assignments for all members of unions,
which could contradict explicitly-made initializations
and thus would produce an unsatisfiable formula for a declaration,
leading to paths being cut off and an unsound analysis.

Because it is quite difficult to fix this completely,
we now disable implicit initializers for unions completely.
Explicit initializers are still handled.

This fixes #255.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@21676 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c