Generifying the AbstractDomain
AbstractDomain and ConfigurableProgram are now both parametrized by the
type of the AbstractState that they use.
This requires less boiler plate code, allows to catch subtle type
errors, and serves as an additional documentation, as the domain has to
indicate over what states does it operate.

The extra required change was introduction of FlatLatticeNoTopDomain.
Before, many analyses were using FlatLatticeDomain with no arguments,
which implies using FlatLatticeDomain.TOPState.
This is potentially confusing, as two different classes of states could
be potentially mixed together.
The introduction of FlatLatticeNoTopDomain aims to enforce the type
safety that one CPA should operate over abstract states of only one
kind.

This change is large, but subsequent generifying changes (Reducer,
TransferRelation, MergeOperator, etc) could be made much smaller,
as the main task (generifying CPA with the type of the state used) is
already done.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/generify_core_interfaces@21757 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c