Replace all calls to CTypeUtils.simplifyType.
Check r22227 for why this is necessary
(the method uses static state and is not thread-safe).

This method is now available in TypeHandlerWithPointerAliasing.
Callers that get types from the CFA use the moved method now.
Other callers (e.g., internal utility methods),
where the type should actually have been simplified already,
now call a different method that does not simplify
but only asserts that a type is already simplified.
This should also be a little bit more efficient.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@22241 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c