ValueAnalysisState returns formula approximation for symbolic values
Use symbolic execution's FormulaCreator to transform symbolic values into formulas.
For this, ValueAnalysisState has to be initialized with different util classes to be able to create the FormulaCreator and FormulaTypeHandler.

Since ValueAnalysisState did not contain much logical classes until now, add the objects as static ones so that not too much changes are necessary.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/error_invariants@22932 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c