New memory model for pointer analysis with regions
Regions are based not only on type, but also on structure field names (so called BnB model).
If the field is not accessed by an address operator then it is placed in a separate region.
This branch is made by refactoring branch pointer_analysis_bnb_model (implemented by Anton Volkov).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/pointer_analysis_bnb_model2@23129 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c