Update to google-java-format 1.1.
This uses a new version of the Google Java Style guide.
The most important change in the new style guide is the import order,
which is now simplified. The IDE configurations are adapted to this.

google-java-format now also reformats Javadoc comments,
making them have a consistent line length and correct Javadoc syntax.
Note that this will destroy some kind of manual formatting in Javadoc
comments like Markdown-style lists.
Please use proper Javadoc syntax instead, which should be used anyway
because all other formatting is already broken when viewing the HTML pages.

Like before, we don't reformat everything with the new formatter
to avoid lots of merge conflicts with existing branches,
but when code is added or changed it should be formatted with
"ant format-diff".

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@23301 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c