BAM: fix for repeated counterexample. PredicateRefiner: adding return value to 'performRefinement'.
In BAM it is possible that one or more CEX appear several times,
and if we notice that, we want to refine the RelevantPredicatesComputer.
Tracking only the last CEX is not sufficient here,
we have cases of swaping traces, where we need to track 'all' of them.

Thanks to Pavel for reporting such a nice example and the basic idea how to deal with it.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@23452 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c