BAM for recursive procedures: bug fix for exit states of main function.
For true-programs we had returned to few exit-states (only new states in
the last iteration of the fixpoint algorithm).
Thus, the reached-set was invalid (root has a successor, that is not in
the reached-set). This causes problems when analysing the reached-set
afterwards, i.e. when using CEX-check-algorithm or in statistics.

For documentation:
As the CEX-checker is applied to an unrolled CEX,
the analysis of a concrete path might be more expensive than the analysis with BAM itself.
An example for this behaviour can be seen for the Fibonacci-tasks (with depth 20),
for which the CEX-checker CBMC has to analyze over 10000 function calls in the unrolled CEX.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@23670 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c