Remove bugs that can be found on http://linuxtesting.org/results/ldv-cpachecker from doc/BugsFound.txt.
The website contains more bugs and is regularly updated,
so we can use it as primary source of bugs found by LDV with CPAchecker,
and keep the list in the file just for those that are not mentioned
online.
Duplicating some (but not all) of the bugs on the website in the file
is just inconsistent.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@24171 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c