ARGPathExporter: More aggressive filter for edges before exporting.
When exporting edges leading to sink-node, we check whether the edge is redundant now, and remove it from the witness.
This maybe leads to smaller witnesses and, more important, avoids some strange behaviour for concurrency witnesses,
where the second node of a started thread is a sink-node and it was merged into the predecessor that should be a non-sink-node on the longer error-path.

An example for such a wrong witness was produced by:
scripts/cpa.sh -valueAnalysis-concurrency test/programs/benchmarks/pthread/fib_bench_false-unreach-call.i

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@24229 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c