ThreadingState: add mapping from threadId to unique identifier.
This mapping might be useful when validating witnesses, because the witness-graphml does only contain an identifier.
Currently we export our threadId 'as is', but there is no specification/guarantee for this and thus we must use a identifier-independent mapping.
Also the automaton only supports integer variables, not identifiers based on strings.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@24244 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c