Change format how bounding conditions (e.g., loop bounds) are represented in the output conditions of conditional model checking.
Currently this uses rational formulas, but this breaks with Princess.
So now we use integers, which are available on all our solvers.

This shouldn't affect anything, because while we encode the bounding
condition in the output condition, we never actually made use of this
(and the current encoding is just a prototype to have something anyway).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@24545 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c