Fix a performance bug introduced in r2887 (yes, 6 years ago).
We want to compute a few of the sets in LoopStructure.Loop lazily
because they are not always needed,
but we ended up re-computing them every time they were accessed.

This mostly affected our termination analysis,
but for code with complex goto loops this could probably significantly slow down
the CFA creation.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@24617 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c