BAM: extract common code from transferrelations into common abstract super-class.
Now the block-start-location is handled differently, no usage of main-block.
We have to update the Refiner, such that there this case is also supported.
The SubtreeRemover will be fixed in the next commit.

As we do no longer start a sub-analysis with the main-block's root-state,
the stack (see default BAMTransferRelation) is empty when analyzing the main-block.
This must be considered when accessing the outerSubtree for expanding abstract states from nested block to main-block.
Otherwise analyses get wrong precisions, for example causing repeated counterexamples in bam-predicate-analysis.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/BAM-parallel@24776 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c