Simplify handling of properties in CPAchecker.
This has two effects:
First, the temporary configuration options like
memsafety.check are not necessary anymore and are removed.
Second, CPAchecker now rejects combinations of properties that are not
supported, like memsafety+termination.
Previously, this was accepted but lead to wrong results
because the configuration for only one of the properties was used.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@24987 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c