change design
Do not store CFABuilder in the builder decorators, because it requires
to copy the decorator to change the builder. Further, it hides the
builder that is changed. Instead pass the builder as parameter to the
append method. This is more transparent.

Make concrete decorators (builder of a specific node) swappable. This
is more flexible, extensible, composable and testable (no mocks
required).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/javascript@25391 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c