Quick fix for exception in termination analysis about instantiating already instantiated variables.
The underlying problem is that TerminationState reports formulas that
areare instantiated, but this does not make sense as a representation of
states. The quick fix is to uninstantiate these formulas, but I suspect
this might indicate a bigger problem in the termination analysis,
because probably these formulas should not be instantiated anyway
(currently, these formulas contain terms like
  (< |main::i@4| |main::i__TERMINATION_PRIMED|),
which looks suspicously like it should have |main::i| instead of |main::i@4|).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@25614 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c