Add an optimization to LoopStructure.findLoops.
Especially in code from real software, included headers can lead to a
large number of declarations in the global scope, each of which will add
one edge at the beginning of the main function.
Because these edges with global declarations form a linear chain,
we know that these edges never participate in a loop,
but our loop-finding algorithm nevertheless analyzed them like regular
edges.
Because this algorithm does not scale well (quadratic for memory,
and worse for time), this can lead to bad performance and crashes due to
out of memory even for moderately sized programs.
By ignoring linear chains of edges at the beginning of functions during
loop detection, we can easily optimize this case.

This fixes #208, where the main function has 44000 nodes,
but only 115 remain after removing the global declarations.
Now loop detection is instantaneous for this file,
previously it failed even with 11 GB of heap space.

Note that unfortunately this does not help for files like those from #307,
which have main functions with more than a million edges with actual code.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@25639 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c