Change the formula encoding of the predicate analysis.
We have certain "variables" (more precisely, nullary UFs)
that represent constant values, for example return values of constant
external functions, or addresses of allocated memory regions.
These variables do not have an SSA index and are not in the SSAMap.
However, when using the method SSAMap.withDefault,
all variables get SSA indices, even those that should not, which is a
bug.

Thus we now mark all variables that should never get a SSA index
with an "@" as suffix, and make sure that we never add indices for such
variables.
This commit also sets this marker for certain kinds of variables
(those mentioned above), but there might be more variables
that should get this marker.

This has the potential to break some code that relies on the internals
of our formula encoding more than it should,
but all code that correctly uses methods such as
FormulaManagerView.parseName will get back the original name (without the marker)
and should not be affected.

In the future, when we are reasonably sure that we have found all
variables that should be marked in this way,
we can probably add more assertions to the index-handling code,
such as that in path formulas there are only variables that have either
an SSA index or such a marker.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@25642 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c