Add a new command-line argument "-benchmark".
This is intended to make benchmarking CPAchecker easier
by replacing a few individual things that benchmarkers previously needed
to care about.

Currently, this option does the following:
- disables Java assertions
- requires users to explicitly specify a heap size with -heap
- disables output files and extended statistics

This also allows to make configurations like -svcomp* more similar
to other configurations in the future: We do not longer need to put
output.disable=true in these configurations because users will be
expected to use -benchmark.
This makes it more convenient to use this configuration in other
circumstances, like debugging.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@25772 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c