Fix a bug in the formula encoding of the predicate analysis.
When we allocate new memory and we assign initial values,
we actually do not need to increment the SSA index,
because the new memory cells were never written before.
This was implemented as an optimization.

However, in the case of a stack allocation of a struct that contains an
array, we attempted to use this optimization (so assigning initial
values with the previous SSA index), but nevertheless incremented the
SSA indices afterwards.
This means that all initial values of structs with arrays on the stack
would have been lost.

However, there was actually another bug in
AssignmentHandler.makeAssignment, which lead to the set updatedRegions
being always empty, such that this bug never occurred.
In r25837 that other bug got accidentally removed
(by avoiding to set updatedRegions to null if it is not null),
and so the current bug now occurs.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@25839 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c