Fix a bug in the array-based heap encoding of the predicate analysis.
With the array-based encoding, we always need to immediately increment
the SSA index for each assignment, and store the new index in the SSAMap
such that another index is used for the next assignment.
Only for the UF-based encoding we can do multiple assignments for
different heap cells with the same index (before adding the retention constraints).

This bug only occurred in specific circumstances when two or more heap
cells were changed in a single assignment, e.g. for struct copies.
It could cause unsoundness and imprecision.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@25851 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c