Enable the array encoding for heap memory in the predicate analysis by default.
This changes the behavior of almost all configurations with a predicate analyis,
except for:
- ldv and ldv-bam, for which this should be evaluated separately
- policy iteration, for which a JavaSMT bug related to Z3 array models occurs
  (config will be changed after this bug is fixed)
- old SV-COMP configs, where we want to keep the existing behavior

All our SMT solvers support arrays, so this change does not add any
restrictions in this regard, but note that array interpolation is
incomplete for all solvers, so we can expect to see failed interpolation
attempts. However, benchmarks show that in many cases where array
interpolation fails, we wouldn't be able to solve the task with the UF
encoding in a timely manner, too.

In addition to chaning the heap encoding,
we now set cpa.predicate.maxArrayLength to unlimited, too.
With the UF encoding, large arrays produce large formulas,
but for the array encoding this is a much smaller problem
(only for initialization of large arrays and for assignments of structs
containing large arrays we will produce large formulas).

Note that this change still does not produce a 100% precise heap
encoding, because for variably-sized arrays we still do not know their
length and have to approximate them (using
cpa.predicate.defaultArrayLength as their length).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@25925 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c