Switch to MathSAT5 as default SMT solver for CPAchecker's predicate analysis.
This is necessary for the upcoming switch to bitprecise encoding as the
default, which is not supported by SMTInterpol.

Configurations that already use Z3 continue to do so.
Furthermore, the configuration kInduction-kidfInvariants
needs a reentrant solver, so we continue using SMTInterpol for this.
For the LDV configurations, this change should be evaluated separately.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@25958 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c