InterpolationStrategies: add new strategy for sequential interpolation.
We can also safely use reversed itps for itp-based analyses like predicate analysis or impact.
ITP(A,B) has same properties than NOT(ITP(B,A)),
but it might look different, depending on the solver's internal strategy.

Example (predicate analysis):
    for (int i = 0; i != 1000000; i++) { __VERIFIER_assert(i <= 1000000); }
With forward-strategy we get ipts like "i=0", "i=1", "i=2",... which unroles the loop completely.
With backward-strategy we get one ipt "i<=1000000" and finish the analysis in the next step.

We can also fall back to the other direction if the solver throws an exception,
because maybe the reverse direction for interpolation succeeds.

For good interpolants we could take a look at
"Duality-Based Interpolation for Quantifier-Free Equalities and Uninterpreted Functions"
from Alt, Hyvarinen, Asadi, and Sharygina.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@25987 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c