Make non-overlapping constraints of memory regions in predicate analysis more precise.
Previously, we completely ignored expressions that we could not
statically evaluate, e.g, arguments to malloc using variables,
or lengths of variably-typed arrays.
In such cases, we used only the configured default array length.
This could lead to overlapping memory regions if the actual allocations
were larger, and thus an imprecise analysis.

Now we ensure that memory regions are always spaced as far apart as
necessary, even for dynamic sizes, by using the size expressions
in appropriate constraints.

This fixes #340.

Note that this does not completely eliminate the use of the configured
default array length, but reduces its importance
(especially for the array encoding).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@25993 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c