improve logging for DelegatingRefiner.
With sv-comp17-config we currently return a wrong result for 'test/programs/benchmarks/recursive
/Primes_true-unreach-call_true-no-overflow_false-termination.c',
because the predicateRefiner aborts after finding a repeated CEX and
the imprecise CEX of the value-analysis is returned without checking it with a proper CEX-check.

With the log-msg we have a chance to look for such cases.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26031 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c