Re-add excluded tasks to most of our integration sets.
Most of these tasks were excluded because they were flapping timeouts,
but since we changed the hosts on which we execute these tasks,
we can re-add them all and then start excluding flapping tasks again.

For some of these test sets, this adds a considerable number of tasks
and increases the used time.
We need to see how this turns out to work,
but maybe because of the resource reduction from the last commit
it works fine. Otherwise we have to remove tasks again.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26038 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c