BAMCPA: change data-structure for storing block abstractions.
This commit applies an idea from branch BAM-COW and
aims for a better identification of block abstractions
when iterating over the ARG after an analysis, e.g. for CEX-check or statistics.
Instead of identifying a used block abstraction by its initial state,
we also use the exit state of the current context.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26052 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c