add method to CPA-utils to extract a CPA or fail directly.
We need this functionality in several places,
and a single implementation might be good,
as the Exception is now also identical across several use-cases.

In some places this commit weakens the assumption about the given CPA,
such that not only WrapperCPAs, but also other kinds of CPA can be used.
The extraction process is defined recursively on the CPAs anyway.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26093 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c