VariableClassification: performance improvement.
The performance problem is caused by "Sets.intersection(smallSet, bigSet).isEmpty()"
(under the assumption that "smallSet" as first parameter is better, as written in Guava-JavaDoc!),
which executes "Collections.disjoint(smallSet, bigSet)",
which then swaps (!) the two parameters, because the first one is instance of "Set".
Thus we were iterating over a growing set, possibly with runtime O(n*n).

Now we directly iterate over the set to avoid such implementation-defined behaviour.
The runtime should be about O(n), depending on number n of variables.

This change should fix #369, but needs to be tested for the benchmarks mentioned in this ticket.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26139 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c