Add handling of a different kind of assumptions
The way the assumptions are instantiated currently in the path formulas
is such that the assumptions hold in the current state. But for
overflow handling it is neccessary to instantiate assumptions before
the edge to the current state is executed. Imagine the assignment:
    x = (x+y) * 0;
For detecting the overflow in the addition, access to the previous
value of x is needed. Thus a new kind of assumptions with different
semantics is added to the interface.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26161 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c