Add a test that ensures that certain program-specific options do not appear in config files.
These options need to be specified by the user depending on the program.
It is only ok to have them in config files for specific use cases,
like LDV or SV-COMP.

In two cases, remove such an option from a config file
(in one of them, the default was used anyway).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26256 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c