A branch for one optimization from ldv-bam
The problem was with dynamic allocation in predicateCPA.
Imagine, we have called two alloc functions. At the analysis stage CPAchecker with BAM
handles the function only once (the second one the result from BAM cache is applied).
At the refinement stage the path with two calls is restored. The alloc identifiers (alloc#2) 
will be different. The obtained predicates will not be applied on the next stage.

The idea is to adjust BAM and not to use BAM cache, when it may lead to wrong results.



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/dynamic-bam-adjustment@26305 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c