Optimize Slicing Abs. to remove more states
States from which the target state(s) is(are) not reachable after
slicing can savely be removed. This saves the effort of splitting and
slicing irrelevant edges later. It should also not interfere with
model checking style predicate analysis:
A successful refinement removes target locations on first encounter
and the removal is designed in a way that states only get removed if
there is at least one target state in the reached set after refinement.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/slicing-abstractions@26400 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c