Optimize Slicing Abs.: avoid splitting if possible
Splitting can sometimes be avoided if we know that the new state
will definitely get removed (all entering edges are infeasible).

This is a shortcut especially for model checking style analysis,
where the abstraction states form a tree. In this case,
Slicing Abstractions (hopefully) reduces more or less to Impact.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/slicing-abstractions@26402 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c