Witness-Export: add label for nodes, such that yEd can show a noce graph.
Tools like yEd do not keep our internal node-ids, but only the 'data' for nodes.
Thus we store the same information twice, as 'node-id' and as 'label'.

yEd allows to show (and layout) witness graphmls with additional data taken from data-tags,
when enabling the "property mapper" feature in yEd.
This might simplify debugging for larger witnesses.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26429 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c