Attempt to configure the Eclipse formatter similarly to google-java-format
We actually do not want to use the Eclipse formatter,
but if a developers does not have the google-java-format plugin
installed, Eclipse automatically falls back to its own formatter
instead of no formatting, so at least we want to have it look similarly.

Also update documentation for Eclipse users related to this.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26471 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c