Remove an assertion for long doubles in AbstractExpressionValueVisitor
This class simply returns the input value in a lot of cases
if it cannot handle it, so we do the same here.
However, I suspect that this is totally broken,
because there are callers like the ExpressionSimplificationVisitor
that assume that an operation like a cast has been handled successfully
if something is returned.

So this class probably needs a complete rewrite from scratch
until we can be sure that it implements the operations precisely.
See also #146.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26581 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c