Update JavaSMT with the new MathSAT 5.5.0-pre1
This brings two independent changes:
- Fix of the "invalid model" error in our termination analysis.
- Use of the reentrant MathSAT version.

The latter finally allows us to provide a bitprecise
kInduction-kidfInvariants configuration.

Furthermore, for consistency with other configurations
I change the solver of kInduction-linear-kidfInvariants from SMTInterpol
to MathSAT5.

We could also consider to make the config kInduction point to
kInduction-kidfInvariants instead of kInduction-dfInvariants,
for consistency with kInduction-linear, which points to
kInduction-linearlinear-kidfInvariants (but this change is not done in
this commit).
The svcomp18 config, which also uses kInduction-dfInvariants, is also
not changed.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26622 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c