Properly implement #340 for variably-sized array types.
In r25993 I forgot to actually take the size expression of
variably-sized array types and use it for the size, although this was
one of the original motivations of #340.

This makes the formula encoding of the Predicate CPA more precise
by preventing illegal overlapping of stack arrays with other
allocations.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26631 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c