Initial implementation of a Newton-based refinement. Currently only supports StrongestPostCondition.
Adds a new Class NewtonRefinementManager that is created and called by the PredicateCPARefiner. 
Stored in PredicateCPARefiner as Optional field . 
Uses the config-option cpa.predicate.refinement.useNewtonRefinement. Does not interfere with previous code as long as option is false.

Tested so far on simple programs with a SBE-BlockStrategy and Princess as Solver.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26632 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c