ARG can be exported to small bitmap using cpa.arg.bitmapGraphic = true.
The first black pixel at the top in the middle represents the root of the ARG.
ARG states are represented by lines.
The distance of a line from the first black pixel represents
the minimum distance of the ARG nodes represented by that line from the
ARG root.
The length of a horizontal line represents the number of ARG states
with the same minimum distance from the ARG root.
Currently, each black pixel represents one ARG state.
The horizontal relation between ARG nodes on different levels
is arbitrary and does not hold any information about their relation
(e.g., whether they are parent and child).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26789 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c