Delete unused component configs, add test for this
Configs that are in the components folder but are not included in any
other configuration are orphaned and therefore deleted. From now on
the buildbot will complain whenever such orphans are created. If you
still need one of the deleted configs, feel free to readd it, but make
sure to either move it out of the components folder or add a config
that actually uses your component.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@26937 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c