Fix a bug in the abstraction cache of the Predicate CPA.
If the SSAMap of the path formula contains indices
which do not appear in the path formula but are relevant for the
predicates, an inappropriate value could be read from the cache
because the indices were not used for the predicates in the cache key.

To the best of our knowledge such a situation can't actually happen
with any standard configurations of the predicate analysis,
but it is nevertheless a bug and should be fixed.

Thanks to Pavel for finding and reporting this!

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/CPALockator-theory@27183 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c