Make CompositeCPA work without a LocationCPA
This is useful e.g. for analysis that treat locations symbolically.
We accumulate the successors by asking the wrapped TransferRelations,
pretty much like in the case where the location is known. This will
currently lead to an exception if the specification uses an automaton,
because the AutomatonTransferRelation does not support the method
getAbstractSuccessor(state, precision).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/slab@27247 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c