Move calculation of violated properties to a reached set
For special errors, like race conditions, violated property related not to a single state,
but to a whole reached set. So, in this case only reached set may say, is there a need in 
refinement procedure or decide what is the result of the launch.
Thus, the functionality is moved to a reached set itself. 

The assertion in CEGAR algorithm is removed due to the similar reason: special analyzes,
which search not the first error, may fail the assertion.
The hacky condition about a size of reached set is not enough for other algorithms, for example, 
those, which do not use a restart algorithm before CEGAR.

The specific reached sets, which use the improved interface, are not yet added.
They will be merged together with new analysis later.



git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27275 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c