ParallelBAM: bugfix to avoid forgotten state updates.
With the change in r27006, we need to re-add states into the waitlist
after an unexpected event like finding a property violation or recursion,
if we want to re-use parts of the ARG later (CEGAR uses cached reachedsets!).
Otherwise we get unsound results, because of insufficient state-space exploration.

Example (compare results with and without Cegar!):
    scripts/cpa.sh \
    -config config/valueAnalysis-parallelBam-Cegar.properties \
    -setprop algorithm.parallelBam.numberOfThreads=1 \
    -setprop cpa.bam.blockHeuristic=FunctionPartitioning \
    -spec ../sv-benchmarks/c/ReachSafety.prp \
    test/programs/benchmarks/product-lines/minepump_spec4_product35_false-unreach-call_false-termination.cil.c

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27303 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c