Remove some irrelevant svn:mergeinfo properties.
SVN supports merges between branches that affect only some paths in the repository,
and this will lead to mergeinfo properties on individual files and directories.
Usually we avoid this, but sometimes it happens by accident.
For the four files from this commit it happened,
but the content of the mergeinfo property is basically the same as for trunk itself,
so it can be removed.

For InvariantsCPA and cwriter such partial merges were done on purpose,
so the mergeinfo property for them needs to kept.


git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27382 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c