Re-implement dominator CPA for CFA nodes.
Dominator CPA previously computed dominators
for LocationStates from a wrapped location CPA.
There's no need to wrap a location CPA,
and no need for a separate Post-Dominator CPA.

Instead, we compute dominators for CFA nodes and rely
on the location CPA in a composite setup:
This way, the dominator state only stores the dominators -
the dominated node is implicitely defined through the location state.
If we want to compute post-dominators, we run a backwards location CPA
(see committed configuration).

We also assume that dominators are computed with BFS traversal,
this allows us to greatly simplify our implementation.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27396 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c