Slicing CPA with two example configs.
Slicing CPA goes between ARGCPA and CompositeCPA
and knows two behaviors:
if the current CFA edge is in the precision
of the slicing CPA, the slicing transfer just delegates to
the composite transfer.
But if the current CFA edge is not in the precision
of the slicing CPA, the slicing transfer delegates a blank edge
with the correct precessor and successor locations to the composite
transfer (to model a noop-operation).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27397 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c