Delegating slicing refiner works with predicate analysis.
Predicate analysis uses the block formulas stored in existing
states for precision refinement and doesn't compute these anew
on the error path - we change this through a new block formula strategy
that really recomputes the block formulas based on partial path
formulas.

In addition, predicate analysis requires a global restart in combination
with slicing, as the block formulas for abstraction will be empty,
otherwise.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27404 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c