LocationCPA Transfer Relation performs a transfer if edge with same predecessor and successor exists.
Edges don't have to be exactly the same, only predecessor and successor
have to be equal.
This allows the slicing CPA to replace any edge with a blank edge
with the same predecessor and successor - if this commit is reverted,
slicing with predicate analysis has to be used without caching of block
formulas.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27408 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c