SlicingCPA slices for a target path for each assume edge and the last edge.
If we only slice for the last edge, the slice is actually
correct (and potentially smaller), but the counterexample
may not be a correct counterexample for the original program.

Assume following program snippet:

```
x = 0;
if (x > 0) {
  x = 1;
}
__VERIFIER_error();
```

The slice will be just `__VERIFIER_error();` because the if-construct
doesn't matter.
But if analysis with slicing finds a counterexample that goes into the
if-branch of the if-construct, the counterexample will be wrong;
Thus,
1. a counterexample check without slicing will proof the target path
  infeasible, but it will be encountered again because the slice
  doesn't care about the if-construct. This creates an infinite loop.
  Or,
2. we perform a counterexample check with slicing, the refinement
  procedure will tell us that the counterexample is feasible,
  but a following cex-check will fail! In addition, the violation
  witness will be invalid.

As a result, we create bigger slices, that don't have these problems
and that can potentially create correct violation witnesses.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27409 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c