Revert r27408, r27481, and r27482.
r27408 was just a temporary change that changes the semantics of the
LocationTransferRelation but should not be applied permanently.
The refactorings based on this are reverted together with this change.

This also re-introduces the quadratic runtime that was improved in
r27482, but we can work on this in the future.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27485 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c