Change scoring for variables to logarithmic scale.
Instead of multiplication we use addition,
which results in much smaller numbers
while keeping comparability of scores.

This avoids a possible overflow for domain-type-based scores in many cases.

Background:
Changing from a scale with values [2,4,16] and MULT towards values [1,2,4] and PLUS
corresponds to a change from linear scale to log scale (or exp scale to linear scale).
The previous approach did not use large integer ranges.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27551 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c