Adjust IntelliJ dependency structure to classpath specified for ant.
If the full `lib` directory is included, as r27604 proposed,
jars with the same name but from different configurations
and with different content may produce problems during compilation
because the wrong jar is picked.

Specifically, this is currently the case with `guava.jar`
in `runtime` and `test` (both the same) and `guava.jar` in
`format-source` (different).

The new structure only includes directories `runtime` and `test`, as
well as libraries in the top level of `lib`.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@27605 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c